+++ test/close-test.js
