+++ test/del-test.js
