+++ test/iterator-no-snapshot-test.js
