+++ test/iterator-snapshot-test.js
