+++ test/iterator-test.js
